+++ examples/use_rust/nrf52840_ble_split/src/central.rs
@@ -92,6 +92,8 @@ async fn main(spawner: Spawner) {
         Some(charging_led),
         false,
         Some(saadc),
+        2000,
+        2806,
     );
     let storage_config = StorageConfig {
         start_addr: 0,