+++ rmk/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add layout macro to! and df!
+- Make nRF52840 voltage divider configurable
 
 ### Changed